Matrices: Add `transform_point`.
I have also added a typedef for point which in theory is a 3x1 matrix,
but in implementation is just a 3 length array.

I have also realized that the matrix multiplication code was flawed if
the output and one of the input were same, so now we do them to a
temporary place and copy after.

For copying I have also implemented a `copy_matrix`, and `copy_point`.